Add command to fetch exchange rates with bean-price
The plugin was lacking a simple way to append current prices to a
ledger. Since we can easily shell out to bean-price in a function, this
is a pretty trivial addition. For convenience, fetch_price() defaults to
fetching today's prices.

Note that for now, we fetch all prices by passing -i to bean-price.